Refactor Members to MemberRecords
